+++ config/routes.rb
@@ -52,6 +52,13 @@ Rails.application.routes.draw do
     post 'concepts/:origin/to_review'   => 'concepts/versions#to_review', :as => 'concept_versions_to_review'
     get 'concepts/:origin/consistency_check' => 'concepts/versions#consistency_check', :as => 'concept_versions_consistency_check'
 
+    post 'collections/:origin/branch'      => 'collections/versions#branch',    :as => 'collection_versions_branch'
+    post 'collections/:origin/merge'       => 'collections/versions#merge',     :as => 'collection_versions_merge'
+    post 'collections/:origin/lock'        => 'collections/versions#lock',      :as => 'collection_versions_lock'
+    post 'collections/:origin/unlock'      => 'collections/versions#unlock',    :as => 'collection_versions_unlock'
+    post 'collections/:origin/to_review'   => 'collections/versions#to_review', :as => 'collection_versions_to_review'
+    get 'collections/:origin/consistency_check' => 'collections/versions#consistency_check', :as => 'collection_versions_consistency_check'
+
     get 'alphabetical_concepts(/:prefix)' => 'concepts/alphabetical#index', :as => 'alphabetical_concepts'
     get 'untranslated_concepts/:prefix'   => 'concepts/untranslated#index', :as => 'untranslated_concepts'
     get 'hierarchical_concepts' => 'concepts/hierarchical#index', :as => 'hierarchical_concepts'