+++ lib/iqvoc/ability.rb
@@ -7,7 +7,7 @@ module Iqvoc
     def initialize(user = nil)
       can :read, Iqvoc::Concept.root_class.instance
       can :read, ::Collection::Base
-      can :read, [::Concept::Base, ::Label::Base], &@@if_published
+      can :read, [::Concept::Base, ::Collection::Base, ::Label::Base], &@@if_published
 
       # static pages
       can :read, :help
@@ -17,18 +17,16 @@ module Iqvoc
         can :destroy, UserSession
 
         if user.owns_role?(:editor) || user.owns_role?(:publisher) || user.owns_role?(:administrator) # Editors and above ...
-          can :manage, ::Collection::Base
-
-          can :read, [::Concept::Base, ::Label::Base]
-          can :create, [::Concept::Base, ::Label::Base]
-          can [:update, :destroy, :unlock], [::Concept::Base, ::Label::Base], :locked_by => user.id, :published_at => nil
-          can :lock, [::Concept::Base, ::Label::Base], :locked_by => nil, :published_at => nil
-          can [:check_consistency, :send_to_review], [::Concept::Base, ::Label::Base], :published_at => nil
-          can :branch, [::Concept::Base, ::Label::Base], &@@if_published
+          can :read, [::Concept::Base, ::Collection::Base, ::Label::Base]
+          can :create, [::Concept::Base, ::Collection::Base, ::Label::Base]
+          can [:update, :destroy, :unlock], [::Concept::Base, ::Collection::Base, ::Label::Base], :locked_by => user.id, :published_at => nil
+          can :lock, [::Concept::Base, ::Collection::Base, ::Label::Base], :locked_by => nil, :published_at => nil
+          can [:check_consistency, :send_to_review], [::Concept::Base, ::Collection::Base, ::Label::Base], :published_at => nil
+          can :branch, [::Concept::Base, ::Collection::Base, ::Label::Base], &@@if_published
         end
 
         if user.owns_role?(:publisher) || user.owns_role?(:administrator) # Publishers and above ...
-          can :merge, [::Concept::Base, ::Label::Base], :published_at => nil
+          can :merge, [::Concept::Base, ::Collection::Base, ::Label::Base], :published_at => nil
         end
 
         if user.owns_role?(:administrator)