+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
   * [Asynchronous](https://github.com/innoq/iqvoc#background-jobs) SKOS exporter
   * Improved web based SKOS importer
+  * Collections are versionable
   * Concepts, collections and concept scheme can be requested as NTriples
   * Rails 4.1
   * Bootstrap 3